Adding ClickHouseException as base Excpetion
