+++ client-v2/src/main/java/com/clickhouse/client/api/ClientException.java
@@ -1,6 +1,6 @@
-package com.clickhouse.client.api.exception;
+package com.clickhouse.client.api;
 
-public class ClientException extends RuntimeException {
+public class ClientException extends ClickHouseException {
 
     public ClientException(String message) {
         super(message);
