+++ client-v2/src/main/java/com/clickhouse/client/api/ClientFaultCause.java
@@ -8,4 +8,5 @@ public enum ClientFaultCause {
     ConnectTimeout,
     ConnectionRequestTimeout,
     SocketTimeout,
+    ServerRetryable,
 }
