+++ client-v2/src/main/java/com/clickhouse/client/api/ClientMisconfigurationException.java
@@ -1,7 +1,5 @@
 package com.clickhouse.client.api;
 
-import com.clickhouse.client.api.exception.ClientException;
-
 /**
  * Represents errors caused by a client misconfiguration.
  */
