+++ client-v2/src/main/java/com/clickhouse/client/api/ConnectionInitiationException.java
@@ -1,12 +1,14 @@
-package com.clickhouse.client.api.exception;
+package com.clickhouse.client.api;
 
-public class ConnectionInitiationException extends ClientException {
+public class ConnectionInitiationException extends ClickHouseException {
 
     public ConnectionInitiationException(String message) {
         super(message);
+        this.isRetryable = true;
     }
 
     public ConnectionInitiationException(String message, Throwable cause) {
         super(message, cause);
+        this.isRetryable = true;
     }
 }
