+++ client-v2/src/main/java/com/clickhouse/client/api/DataTransferException.java
@@ -1,0 +1,12 @@
+package com.clickhouse.client.api;
+
+public class DataTransferException extends ClickHouseException {
+
+    public DataTransferException(String message) {
+        super(message);
+    }
+
+    public DataTransferException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
