+++ client-v2/src/main/java/com/clickhouse/client/api/metadata/NoSuchColumnException.java
@@ -1,6 +1,6 @@
 package com.clickhouse.client.api.metadata;
 
-import com.clickhouse.client.api.exception.ClientException;
+import com.clickhouse.client.api.ClientException;
 
 public class NoSuchColumnException extends ClientException {
