+++ client-v2/src/main/java/com/clickhouse/client/api/query/NullValueException.java
@@ -1,6 +1,6 @@
 package com.clickhouse.client.api.query;
 
-import com.clickhouse.client.api.exception.ClientException;
+import com.clickhouse.client.api.ClientException;
 
 /**
  * Throw when a null value cannot be returned because of data type.
