+++ client-v2/src/test/java/com/clickhouse/client/ProxyTests.java
@@ -1,7 +1,7 @@
 package com.clickhouse.client;
 
 import com.clickhouse.client.api.Client;
-import com.clickhouse.client.api.exception.ClientException;
+import com.clickhouse.client.api.ClientException;
 import com.clickhouse.client.api.ClientMisconfigurationException;
 import com.clickhouse.client.api.enums.Protocol;
 import com.clickhouse.client.api.enums.ProxyType;
